Move finality tests into `tests.rs`
